+++ src/SMAPI/Framework/ContentManagers/GameContentManagerForAssetPropagation.cs
@@ -25,9 +25,9 @@ namespace StardewModdingAPI.Framework.ContentManagers
             : base(name, serviceProvider, rootDirectory, currentCulture, coordinator, monitor, reflection, onDisposing, onLoadingFirstAsset, onAssetLoaded, aggressiveMemoryOptimizations) { }
 
         /// <inheritdoc />
-        public override T Load<T>(IAssetName assetName, LanguageCode language, bool useCache)
+        public override T LoadExact<T>(IAssetName assetName, bool useCache)
         {
-            T data = base.Load<T>(assetName, language, useCache);
+            T data = base.LoadExact<T>(assetName, useCache);
 
             if (data is Texture2D texture)
                 texture.Tag = this.Tag;